intial revisit of tracking SORT
