+++ crabs/detection_tracking/inference_model.py
@@ -7,6 +7,7 @@ import numpy as np
 import torch
 import torchvision.transforms as transforms
 from sort import Sort
+from crabs.detection_tracking.detection_utils import drawing_detection
 
 # select device (whether GPU or CPU)
 device = (
@@ -59,163 +60,49 @@ class Detector_Inference:
         # TODO: deal with different model
         self.trained_model = torch.load(
             self.args.model_dir,
-            # map_location=torch.device("cpu")
+            map_location=torch.device("cpu")
         )
 
-    def apply_grayscale_and_blur(
-        self,
-        frame: np.array,
-        kernel_size: list,
-        sigmax: int,
-    ) -> np.array:
-        """
-        Convert the frame to grayscale and apply Gaussian blurring.
+    # def __inference(self, frame: np.ndarray) -> np.ndarray:
+    #     """
+    #     Perform inference on a single frame of the video.
 
-        Parameters
-        ----------
-        frame : np.array
-            frame array read from the video capture
-        kernel_size : list
-            kernel size for GaussianBlur
-        sigmax : int
-            Standard deviation in the X direction of the Gaussian kernel
+    #     Args:
+    #         frame (np.ndarray): The input frame as a NumPy array.
 
-        Returns
-        -------
-        gray_frame : np.array
-            grayscaled input frame
-        blurred_frame : np.array
-            Gaussian-blurred grayscaled input frame
-        """
-        # convert the frame to grayscale frame
-        gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-        # apply Gaussian blurring
-        blurred_frame = cv2.GaussianBlur(gray_frame, kernel_size, sigmax)
+    #     Returns:
+    #         None
+    #     """
+    #     self.trained_model.eval()
+    #     transform = transforms.Compose(
+    #         [
+    #             transforms.ToTensor(),
+    #         ]
+    #     )
+    #     img = transform(frame)
+    #     img = img.to(device)
 
-        return gray_frame, blurred_frame
+    #     img = img.unsqueeze(0)
+    #     prediction = self.trained_model(img)
+    #     pred_score = list(prediction[0]["scores"].detach().cpu().numpy())
 
-    def compute_background_subtracted_frame(
-        self,
-        blurred_frame,
-        mean_blurred_frame,
-        max_abs_blurred_frame,
-    ):
-        """
-        Compute the background subtracted frame for the
-        input blurred frame, given the mean and max absolute frames of
-        its corresponding video.
+    #     # if not self.args.sort:
+    #     #     from inference import inference_detection
 
-        Parameters
-        ----------
-        blurred_frame : np.array
-            Gaussian-blurred grayscaled input frame
-        mean_blurred_frame : np.array
-            mean of all blurred frames in the video
-        max_abs_blurred_frame : np.array
-            pixelwise max absolute value across all blurred frames in the video
+    #     #     frame_out = inference_detection(
+    #     #         frame, prediction, pred_score, self.score_threshold
+    #     #     )
 
-        Returns
-        -------
-        background_subtracted_frame : np.array
-            normalised difference between the blurred frame f and
-            the mean blurred frame
-        """
-        return (
-            ((blurred_frame - mean_blurred_frame) / max_abs_blurred_frame) + 1
-        ) / 2
+    #     # else:
+    #     # from inference import inference_tracking
 
-    def compute_motion_frame(
-        self,
-        frame_delta,
-        background_subtracted_frame,
-        mean_blurred_frame,
-        max_abs_blurred_frame,
-    ):
-        """
-        _summary_.
-
-        Parameters
-        ----------
-        frame_delta : int
-            difference in number of frames used to compute the motion
-            channel is computed
-        background_subtracted_frame : np.array
-            normalised difference between the blurred frame f and
-            the mean blurred frame
-        mean_blurred_frame : np.array
-            mean of all blurred frames in the video
-        max_abs_blurred_frame : np.array
-            pixelwise max absolute value across all blurred frames in the video
-
-        Returns
-        -------
-        motion_frame : np.array
-            absolute difference between the background subtracted frame f
-            and the background subtracted frame f+delta
-        """
-        # compute the blurred frame frame_idx+delta
-        _, blurred_frame_delta = self.apply_grayscale_and_blur(
-            frame_delta,
-            [5, 5],
-            0,
-        )
-        # compute the background subtracted for frame_idx + delta
-        background_subtracted_frame_delta = (
-            self.compute_background_subtracted_frame(
-                blurred_frame_delta,
-                mean_blurred_frame,
-                max_abs_blurred_frame,
-            )
-        )
-
-        # compute the motion channel for frame_idx
-        return np.abs(
-            background_subtracted_frame_delta - background_subtracted_frame,
-        )
-
-    def __inference(
-        self, final_frame: np.ndarray, frame: np.ndarray
-    ) -> np.ndarray:
-        """
-        Perform inference on a single frame of the video.
-
-        Args:
-            frame (np.ndarray): The input frame as a NumPy array.
-
-        Returns:
-            None
-        """
-        self.trained_model.eval()
-        transform = transforms.Compose(
-            [
-                transforms.ToTensor(),
-            ]
-        )
-        img = transform(final_frame)
-        img = img.to(device)
-
-        img = img.unsqueeze(0)
-        prediction = self.trained_model(img)
-        pred_score = list(prediction[0]["scores"].detach().cpu().numpy())
-
-        if not self.args.sort:
-            from _inference import inference_detection
-
-            frame_out = inference_detection(
-                frame, prediction, pred_score, self.score_threshold
-            )
-
-        else:
-            from _inference import inference_tracking
-
-            frame_out = inference_tracking(
-                frame,
-                prediction,
-                pred_score,
-                self.score_threshold,
-                self.sort_crab,
-            )
-        return frame_out
+    #     frame_out = inference_tracking(
+    #         frame,
+    #         prediction,
+    #         pred_score,
+    #         self.score_threshold,
+    #     )
+    #     return frame_out
 
     def _load_video(self) -> None:
         """
@@ -240,91 +127,28 @@ class Detector_Inference:
             output_file, output_codec, cap_fps, (frame_width, frame_height)
         )
 
-        if args.image_stack:
-            frame_count = 0
-
-            # get image size
-            width = video.get(cv2.CAP_PROP_FRAME_WIDTH)
-            height = video.get(cv2.CAP_PROP_FRAME_HEIGHT)
-
-            # initialise array for mean blurred frame
-            mean_blurred_frame = np.zeros((int(height), int(width)))
-
-            # initialise array for max_abs_blurred_frame
-            max_abs_blurred_frame = np.zeros((int(height), int(width)))
-
-            while video.isOpened() and frame_count <= 500:
-                ret, frame = video.read()
-                if not ret:
-                    print("No frame read. Exiting...")
-                    break
-
-                else:
-                    # Apply transformations to the frame
-                    gray_frame, blurred_frame = self.apply_grayscale_and_blur(
-                        frame, [5, 5], 0
-                    )
-
-                    # accumulate blurred frames
-                    mean_blurred_frame += blurred_frame
-
-                    # accumulate max absolute values
-                    max_abs_blurred_frame = np.maximum(
-                        max_abs_blurred_frame, abs(blurred_frame)
-                    )
-
-                    frame_count += 1
-            # compute the mean
-            mean_blurred_frame = mean_blurred_frame / frame_count
-            print(mean_blurred_frame.shape)
-            # video.release()
-
+        self.trained_model.eval()
+        transform = transforms.Compose(
+            [
+                transforms.ToTensor(),
+            ]
+        )
+        
         while video.isOpened():
             ret, frame = video.read()
             if not ret:
                 print("No frame read. Exiting...")
                 break
 
-            if args.image_stack:
-                gray_frame, blurred_frame = self.apply_grayscale_and_blur(
-                    frame, [5, 5], 0
-                )
+            # frame_out = self.__inference(frame)
+            img = transform(frame)
+            img = img.to(device)
 
-                background_subtracted_frame = (
-                    self.compute_background_subtracted_frame(
-                        blurred_frame,
-                        mean_blurred_frame,
-                        max_abs_blurred_frame,
-                    )
-                )
-
-                # Compute motion channel using the updated mean and max_abs values
-                video.set(
-                    cv2.CAP_PROP_POS_FRAMES,
-                    video.get(cv2.CAP_PROP_POS_FRAMES) + 10,
-                )
-                success_delta, frame_delta = video.read()
-                if not success_delta:
-                    print("Cannot read frame. Exiting...")
-                    break
-
-                motion_frame = self.compute_motion_frame(
-                    frame_delta,
-                    background_subtracted_frame,
-                    mean_blurred_frame,
-                    max_abs_blurred_frame,
-                )
-
-                # Stack the channels
-                final_frame = np.dstack(
-                    [gray_frame, background_subtracted_frame, motion_frame]
-                ).astype(np.float32)
-                final_frame = (final_frame * 255).astype(np.uint8)
-
-            else:
-                final_frame = frame
-
-            frame_out = self.__inference(final_frame, frame)
+            img = img.unsqueeze(0)
+            prediction = self.trained_model(img)
+            frame_out = drawing_detection(
+                img, prediction, score_threshold=0.5
+            )
             out.write(frame_out)
 
             cv2.imshow("frame", frame_out)
@@ -379,14 +203,8 @@ if __name__ == "__main__":
     parser.add_argument(
         "--sort",
         type=bool,
-        default=False,
+        default=True,
         help="running sort as tracker",
-    )
-    parser.add_argument(
-        "--image_stack",
-        type=bool,
-        default=False,
-        help="using additional channels images as the input",
     )
 
     args = parser.parse_args()
