+++ crabs/detection_tracking/train_model.py
@@ -2,14 +2,13 @@ import argparse
 import json
 
 import torch
+import pytorch_lightning as pl
 import yaml  # type: ignore
-from detection_utils import (
-    create_dataloader,
-    get_train_transform,
-    myFasterRCNNDataset,
+from crabs.detection_tracking.detection_utils import (
+    load_dataset,
     save_model,
 )
-from models import create_faster_rcnn, train_faster_rcnn
+from crabs.detection_tracking.models import FasterRCNN
 
 
 class Dectector_Train:
@@ -30,7 +29,7 @@ class Dectector_Train:
         The model use to train the detector.
     """
 
-    def __init__(self, args: argparse.Namespace):
+    def __init__(self, args):
         self.config_file = args.config_file
         self.main_dir = args.main_dir
         self.annotation_file = args.annotation_file
@@ -38,75 +37,24 @@ class Dectector_Train:
         self.annotation = f"{self.main_dir}/annotations/{self.annotation_file}"
         self.load_config_yaml()
 
-    def load_config_yaml(self) -> None:
-        """Load a YAML file describing the training setup"""
-
+    def load_config_yaml(self):
         with open(self.config_file, "r") as f:
-            self.config = yaml.safe_load(f)  # type: dict
-        print(self.config)
+            self.config = yaml.safe_load(f)
 
-    def _load_dataset(self) -> None:
-        """Load images and annotation file for training"""
-
-        with open(self.annotation) as json_file:
-            coco_data = json.load(json_file)
-
-        self.train_file_paths = []
-        for image_info in coco_data["images"]:
-            image_id = image_info["id"]
-            image_id -= 1
-            image_file = image_info["file_name"]
-            video_file = image_file.split("_")[1]
-
-            if (
-                video_file == "09.08.2023-03-Left"
-                or video_file == "10.08.2023-01-Left"
-                or video_file == "10.08.2023-01-Right"
-            ):
-                continue
-
-            # taking the first 40 frames per video as training data
-            if image_id % 50 < 40:
-                self.train_file_paths.append(image_file)
-
-        self.train_dataset = myFasterRCNNDataset(
+    def train_model(self):
+        train_dataloader = load_dataset(
             self.main_dir,
-            self.train_file_paths,
             self.annotation,
-            transforms=get_train_transform(),
+            self.config["batch_size"],
+            training=True,
         )
+        lightning_model = FasterRCNN(self.config)
 
-        self.train_dataloader = create_dataloader(
-            self.train_dataset, self.config["batch_size"]
-        )
+        trainer = pl.Trainer(max_epochs=self.config["num_epochs"], accelerator="cpu")
 
-    def train_model(self) -> None:
-        """Train the model using the provided configuration"""
-
-        self._load_dataset()
-        # select GPU if available
-        device = (
-            torch.device("cuda")
-            if torch.cuda.is_available()
-            else torch.device("cpu")
-        )
-
-        # create model
-        self.model = create_faster_rcnn(num_classes=self.config["num_classes"])
-        self.model.to(device)
-
-        optimizer = torch.optim.Adam(
-            self.model.parameters(),
-            lr=self.config["learning_rate"],
-            weight_decay=self.config["wdecay"],
-        )
-
-        trained_model = train_faster_rcnn(
-            self.config, self.model, self.train_dataloader, optimizer
-        )
-
+        trainer.fit(lightning_model, train_dataloader)
         if self.config["save"]:
-            save_model(trained_model)
+            save_model(lightning_model)
 
 
 def main(args) -> None:
@@ -131,7 +79,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--config_file",
         type=str,
-        default="config/faster_rcnn.yaml",
+        default="crabs/detection_tracking/config/faster_rcnn.yaml",
         help="location of YAML config to control training",
     )
     parser.add_argument(
@@ -144,7 +92,7 @@ if __name__ == "__main__":
         "--model_name",
         type=str,
         default="faster_rcnn",
-        help="the model to use to train the object detection. Options: faster_rcnn",
+        help="the model to use to train the object detection.",
     )
     parser.add_argument(
         "--annotation_file",
